pyproject.toml
[project]
name = "ToolsBot"
version = "0.1.0"
description = "ToolsBot"
readme = "README.md"
requires-python = ">=3.10, <4.0"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[tool.poetry.dependencies]
annotated-types = "0.7.0"
anyio = "4.10.0"
argcomplete = "3.6.2"
asgiref = "3.10.0"
asyncio-dgram = "2.2.0"
base58 = "2.1.1"
certifi = "2025.8.3"
cffi = "2.0.0"
charset-normalizer = "3.4.3"
click = "8.2.1"
colorama = "0.4.6"
construct = "2.10.70"
cryptography = "44.0.3"
dnspython = "2.8.0"
exceptiongroup = "1.3.0"
fastapi = "0.116.1"
h11 = "0.16.0"
httptools = "0.6.4"
idna = "3.10"
keyboard = "0.13.5"
linkify-it-py = "2.0.3"
loguru = "0.7.3"
markdown-it-py = "4.0.0"
mcstatus = "12.0.5"
mdit-py-plugins = "0.5.0"
mdurl = "0.1.2"
mouseinfo = "0.1.3"
msgpack = "1.1.1"
multidict = "6.6.4"
nonebot-adapter-onebot = "2.4.6"
nonebot-adapter-qq = "1.6.5"
nonebot2 = "2.4.3"
nonechat = "0.6.1"
nuitka = "2.7.14"
ordered-set = "4.1.0"
packaging = "25.0"
pillow = "11.3.0"
pipx = "1.7.1"
platformdirs = "4.4.0"
propcache = "0.3.2"
pyautogui = "0.9.54"
pycparser = "2.22"
pycryptodome = "3.23.0"
pydantic = "2.11.7"
pydantic-core = "2.33.2"
pygetwindow = "0.0.9"
pygments = "2.19.2"
pygtrie = "2.5.0"
pymsgbox = "2.0.1"
pyperclip = "1.11.0"
pyrect = "0.2.0"
pyscreeze = "1.0.1"
python-dotenv = "1.1.1"
pytweening = "1.2.0"
pyyaml = "6.0.2"
requests = "2.32.5"
rich = "14.1.0"
schedule = "1.2.2"
setuptools = "80.9.0"
sniffio = "1.3.1"
sqlparse = "0.5.3"
starlette = "0.47.3"
textual = "3.7.1"
typing-inspection = "0.4.1"
typing-extensions = "4.15.0"
tzdata = "2025.2"
uc-micro-py = "1.0.3"
urllib3 = "2.5.0"
userpath = "1.9.2"
uvicorn = "0.35.0"
watchfiles = "1.1.0"
websockets = "15.0.1"
win32-setctime = "1.2.0"
yarl = "1.20.1"
zstandard = "0.24.0"
lxml = ">=6.0.2"
isort = "^7.0.0"
[tool.nonebot]
adapters = [
    {name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11"},
]
plugins = []
plugin_dirs = ["plugins"]
builtin_plugins = []
